Helics interface now implemented in a seperate module.
